# sync-branch/server/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SyncBranch API",
    "description": "API documentation for SyncBranch project.",
    "version": "1.0.0"
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http", "https"],
  "paths": {
...
    "/profile/update": {
      "put": {
        "summary": "Updates the profile of the authenticated user",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "The new email of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "new_email@example.com"
            }
          },
          {
            "name": "spotify_user_id",
            "in": "body",
            "description": "The new Spotify user ID.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "new_spotify_id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Profile updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/messaging/send": {
      "post": {
        "summary": "Sends a message to another user",
        "parameters": [
          {
            "name": "recipient",
            "in": "body",
            "description": "The recipient's username",
            "required": true,
            "schema": {
              "type": "string",
              "example": "recipient_user"
            }
          },
          {
            "name": "message",
            "in": "body",
            "description": "The message content",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Hello!"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message sent successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Message sent successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/messaging/retrieve": {
      "get": {
        "summary": "Retrieves messages for the authenticated user",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "type": "integer",
            "default": 1,
            "example": 1
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Messages per page",
            "required": false,
            "type": "integer",
            "default": 10,
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "messages": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "content": {
                        "type": "string",
                        "example": "Hello!"
                      },
                      "sender": {
                        "type": "string",
                        "example": "example_user"
                      },
                      "sent_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Bearer token"
    },
    "BasicAuth": {
      "type": "basic",
      "description": "Basic Authentication with username and password"
    }
  },
  "security": [
    {
      "Bearer": []
    },
    {
      "BasicAuth": []
    }
  ]
}